Refactor(BridgeTest,ControlTest,MoveTest):Validation 폴더로 이동
