+++ __tests__/Validation/BridgeTest.js
@@ -1,4 +1,4 @@
-const BridgeValidation = require('../src/Validation/BridgeValidation');
+const BridgeValidation = require('../../src/Validation/BridgeValidation');
 
 describe('다리 생성 테스트', () => {
   test('입력이 3이상 20이하가 아닌경우', () => {
